Fix 2% flake.
It turns out that there are two distinct port spaces in Linux: IPv4 and
IPv6. That is, requesting port 0 for an IPv4 address may give you a port
already allocated to an IPv6 address on the system. However, depending
on the resolver used, 'localhost' can resolve to either 127.0.0.1 or
::1.

In the average case, there is no IPv6 address listening on the same
port, so the IPv6 port passes the packets on to IPv4, where it ends up
with the TcpProxy. However, it turns out that the Bazel daemon runs its
own gRPC server, listening on a kernel-allocated port, which it uses to
communicate with the bazel command line client. In around 2% of cases,
the TcpProxy will be allocated the same port as the Bazel gRPC server
and the channel will communicate with *it* instead of the intended
TcpProxy. This results in an UNIMPLEMENTED error, since different
methods are implemented on the server.

In principle, the cross talk could have happened with any service
listening on ::1, but this case was particularly pernicious, because the
error message did not clearly point to cross talk.